@@ -1,5 +1,5 @@
 {
-  "name": "pasona",
+  "name": "anichur",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,41 +8,17 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "add": "^2.0.6",
-    "axios": "^0.21.4",
-    "bootstrap": "^4.4.1",
-    "bootstrap-vue": "^2.4.0",
-    "core-js": "^3.4.4",
-    "feather-icons": "^4.26.0",
-    "jquery": "^3.4.1",
-    "jw-vue-pagination": "^1.0.3",
-    "moment": "^2.29.1",
-    "odometer": "^0.4.8",
-    "pwa": "^1.9.7",
-    "v-wow": "git+https://github.com/guisoares2011/vue-wow.git",
-    "vue": "^2.6.14",
-    "vue-backtotop": "^1.6.1",
-    "vue-carousel": "^0.18.0",
-    "vue-clipboard2": "^0.3.1",
-    "vue-cool-lightbox": "^1.0.10",
-    "vue-faq-accordion": "^1.3.0",
-    "vue-feather": "^1.0.0",
-    "vue-odometer": "^1.0.2",
-    "vue-router": "^3.1.5",
-    "vue-slick": "^1.1.15",
-    "vue-toasted": "^1.1.27",
-    "vuex": "^3.1.2"
+    "core-js": "^3.8.3",
+    "vue": "^3.2.13"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.1.0",
-    "@vue/cli-plugin-eslint": "^4.1.0",
-    "@vue/cli-service": "^4.1.0",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "node-sass": "^4.13.1",
-    "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.14"
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3"
   },
   "eslintConfig": {
     "root": true,
@@ -50,19 +26,18 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
-    "rules": {
-      "no-console": "off",
-      "no-mixed-spaces-and-tabs": 0
-    },
     "parserOptions": {
-      "parser": "babel-eslint"
-    }
+      "parser": "@babel/eslint-parser"
+    },
+    "rules": {}
   },
   "browserslist": [
     "> 1%",
-    "last 2 versions"
+    "last 2 versions",
+    "not dead",
+    "not ie 11"
   ]
 }
